requirements.txt
discord.py
aiohttp
pystyle